get rid of more deprecation warnings in OnlineDB
