--- OnlineDB/SiStripConfigDb/test/plugins/module.cc
@@ -1,5 +1,0 @@
-
-#include "FWCore/Framework/interface/MakerMacros.h"
-
-#include "OnlineDB/SiStripConfigDb/test/plugins/testSiStripConfigDb.h"
-DEFINE_FWK_MODULE(testSiStripConfigDb);
